requirements.txt
streamlit
google-generativeai
pandas
requests
beautifulsoup4
matplotlib
python-dotenv